test saveAll behavior
Found another wrong thinking on the expected behavior. The initial goal of this test was to see the behavior of repository if I insert two books with same ISBN, which is @Id and should be unique. I decided to use saveAll to save the books via one call, wrapping them into Flux. First of all - there was no error whatsoever. I examined the returned Flux and initially came to conclusion that this is because books from input are processed separately and therefore the first one gets saved first, then the second one is treated like an update to the first. This was supported by the findAll() call which was yielding 1 item. I attempted to add an assertion about that, i.e. expect book1, then book2 in the verification of saveAll() result, and expecting book2 in verification of findAll() result. It appeared to work from first sight, but later I found that the test is not stable. I.e. sometimes, findAll() would return book1 instead of book2. And saveAll() would return book2 before book1.

After assessing of the situation, I found that reactive saveAll() works differently. The input elements are processed in parallel manner on a thread pool. I.e. in my test I actually emulate race condition of two users saving same entity with different content. Actually, this should give error as I would expect, but looks like only after I set up optimistic locking and introduce "version" field to my Book document. Without it, the order in which the data will be written to the database is not predictable. So the only assertion I can make here is that with two input elements to saveAll(), it will return two outputs in the result. I played with delay between appearance of inputs in Flux using delayElements() with period enough for the books to be written in expected order, but this is just to better understand the issue.